Apply ascending order for execution time array
Apply the ascending order for the local variable 'exec_times'
by using quick sort.

This can make the option 'simulation' pass the test (Execution
Time is the constant time) for the function 'q_insert_tail' and
'q_size'.

This commit imitates AdrianHuang's repo to learn how to write
good commit message, and the workflow of development.

ref: https://github.com/AdrianHuang/lab0-c.git